refactor: onboarding 관련 쿼리 커스텀 훅 제거
