+++ apps/client/src/pages/onboarding/components/artist-search.tsx
@@ -1,8 +1,10 @@
+import { useQuery } from '@tanstack/react-query';
+
 import { SearchBar, SearchSuggestionList } from '@confeti/design-system';
+import { ARTIST_RELATED_QUERY_OPTIONS } from '@shared/apis/onboard/artist-related-queries';
 import { useDebouncedKeyword } from '@shared/hooks/use-debounce-keyword';
 
 import { ONBOARD_LIMIT } from '../constants/limit';
-import { useArtistRelatedKeyword } from '../hooks/use-onboard';
 
 import * as styles from './artist-search.css';
 
@@ -18,10 +20,12 @@ const ArtistSearch = ({
   const { keyword, debouncedKeyword, handleInputChange } =
     useDebouncedKeyword();
 
-  const relatedKeywordsData = useArtistRelatedKeyword({
-    keyword: debouncedKeyword,
+  const { data: relatedKeywordsData } = useQuery({
+    ...ARTIST_RELATED_QUERY_OPTIONS.KEYWORD(
+      debouncedKeyword,
+      ONBOARD_LIMIT.RELATED_KEYWORD,
+    ),
     enabled: !!debouncedKeyword.trim(),
-    limit: ONBOARD_LIMIT.RELATED_KEYWORD,
   });
 
   const hasArtistResults = (relatedKeywordsData?.artists?.length ?? 0) > 0;
